+++ packages/ui/src/v1/components/code-editor/codemirror/utils.ts
@@ -23,3 +23,24 @@ export const replaceSourceIfNeeded = (view: EditorView, value: string) => {
     }),
   );
 };
+
+export type CancelGroup = {
+  add: (fn: () => void) => void;
+  cancel: () => void;
+};
+
+export function createCancelGroup(): CancelGroup {
+  const callbacks: Array<() => void> = [];
+  
+  return {
+    add: (fn: () => void) => {
+      callbacks.push(fn);
+    },
+    cancel: () => {
+      for (const fn of callbacks) {
+        fn();
+      }
+      callbacks.length = 0;
+    }
+  };
+}
