--- packages/ui/src/v1/shared/cancel.ts
@@ -1,20 +1,0 @@
-export type CancelGroup = {
-  add: (fn: () => void) => void;
-  cancel: () => void;
-};
-
-export function createCancelGroup(): CancelGroup {
-  const callbacks: Array<() => void> = [];
-  
-  return {
-    add: (fn: () => void) => {
-      callbacks.push(fn);
-    },
-    cancel: () => {
-      for (const fn of callbacks) {
-        fn();
-      }
-      callbacks.length = 0;
-    }
-  };
-}
